Add custom errors, write parse_play_by_play.py
The `errors` module contains custom exception types.

The `data_helpers` module contains classes and objects that help verify
both raw data and JSON data. Currently it contains a set of the team
codes to allow checking if team codes are valid.

Added `parse_play_by_play.py`: It contains a function to convert field
position in the raw data to distance to goal line, and a function to
convert the game clock to seconds since the game started. Unittests for
these functions are included, and they all pass.