Move dictionary declaration in converter.py
Dictionaries are now declared in __init_json(), instead of in each
function that fills them. Where a dictionary is completely generated
from one function (for example, weather), or doesn't make much sense to
have ahead of time (officials), then it is left as None and filled in by
the parsing functions.