Change how converter.py saves files by default
Converter now only saves files if they have changed in a way other than
the _version hash. If the version has changed, but the data hasn't
actually changed, then we don't care. This will make it easier to see
meaningful changes in the data git repository.

This default behavior can be overridden with the new "--force-overwrite"
command line flag.