Rework numbering of plays
Using the number that came with the raw data led to play numbering
starting at 0, 1, or even 2.  Play numbering now always starts at 0, and
counts up for each play, even if it fails to parse. This allows gaps in
the numbering to be used to detect certain types of parser failures.